requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
astor=0.8.1=pypi_0
attrdict=2.0.1=pypi_0
babel=2.14.0=pypi_0
bce-python-sdk=0.9.5=pypi_0
beautifulsoup4=4.12.3=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.3=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.1=pypi_0
click=8.1.7=pypi_0
contourpy=1.1.1=pypi_0
cssselect=1.2.0=pypi_0
cssutils=2.9.0=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.9=pypi_0
decorator=5.1.1=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.110.0=pypi_0
fire=0.6.0=pypi_0
flask=3.0.2=pypi_0
flask-babel=4.0.0=pypi_0
fonttools=4.43.1=pypi_0
future=1.0.0=pypi_0
h11=0.14.0=pypi_0
httptools=0.6.1=pypi_0
idna=3.4=pypi_0
imageio=2.34.0=pypi_0
imgaug=0.4.0=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.3=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_2
libgomp=13.2.0=h807b86a_2
libnsl=2.0.1=hd590300_0
libsqlite=3.43.2=h2797004_0
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=hd590300_5
lmdb=1.4.1=pypi_0
lxml=5.1.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.0=pypi_0
ncurses=6.4=h59595ed_1
networkx=3.2.1=pypi_0
numpy=1.26.1=pypi_0
onnx=1.15.0=pypi_0
opencv-contrib-python=4.6.0.66=pypi_0
opencv-python=4.6.0.66=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.1.4=hd590300_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
paddle-bfloat=0.1.7=pypi_0
paddleocr=2.7.0.3=pypi_0
paddlepaddle-gpu=2.4.2=pypi_0
pandas=2.1.2=pypi_0
pdf2docx=0.5.8=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
premailer=3.10.0=pypi_0
protobuf=3.20.0=pypi_0
psutil=5.9.6=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyclipper=1.3.0.post5=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.6.3=pypi_0
pydantic-core=2.16.3=pypi_0
pymupdf=1.20.2=pypi_0
pyparsing=3.1.1=pypi_0
python=3.10.0=h543edf9_3_cpython
python-dateutil=2.8.2=pypi_0
python-docx=1.1.0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
rapidfuzz=3.6.2=pypi_0
rarfile=4.1=pypi_0
readline=8.2=h8228510_1
requests=2.31.0=pypi_0
scikit-image=0.22.0=pypi_0
scipy=1.11.3=pypi_0
seaborn=0.13.0=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
shapely=2.0.3=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.43.2=h2c6b66d_0
starlette=0.36.3=pypi_0
termcolor=2.4.0=pypi_0
thop=0.1.1-2209072238=pypi_0
tifffile=2024.2.12=pypi_0
tk=8.6.13=h2797004_0
torch=1.12.0+cu102=pypi_0
torchaudio=0.12.0+cu102=pypi_0
torchvision=0.13.0+cu102=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023.3=pypi_0
ultralytics=8.0.203=pypi_0
urllib3=2.0.7=pypi_0
uvicorn=0.28.0=pypi_0
uvloop=0.19.0=pypi_0
visualdl=2.5.3=pypi_0
watchfiles=0.21.0=pypi_0
websockets=12.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.3=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
